Move and rename AnnotationSolrQueryConverter to FilterOnlyAnnotationQueryConverter.
